chg: fallbacks default admin user to `root` if current user is not found
It's better to assume that if the current user can't be found it's because we are in a strip-down container and the user is root.

`admin` is also no longer hardcoded in server logs

Integration tests failed in previous commit because of the use of the pure Go implementation of os/user (`osusergo`) in the docker build which allowed to find the current user in the distroless images without libc. The integrations tests where assuming default user is always the fallback one (`admin`)